Change how messages are broadcast
Massages with the same to and from address are sent from the address to it,
instead of being broadcast. To send a broadcast the user includes
"broadcast" anywhere in the to address.

This allows users to use the send to chan method in compliance with the new
chan support in PyBitmessage.