Issue #12: Allow per-address fetches
Add "currentAddress" to the bminterface as a global variable
Change handleUser to set the currentAddress if username begins with "BM-"
This change is meant to allow different users to access their bitmessages
Note this will not work concurrently! once USER is sent, all subsequent messages assume the same user until a different USER is defined.

This closes issue #12